app : rename detail screen component
