+++ app/lib/ui/components/component_detail_screen.dart
